Move InputSocket and OutputSocket in types package
