+++ haystack/core/component/__init__.py
@@ -1,7 +1,7 @@
 # SPDX-FileCopyrightText: 2022-present deepset GmbH <info@deepset.ai>
 #
 # SPDX-License-Identifier: Apache-2.0
-from haystack.core.component.component import component, Component
-from haystack.core.component.sockets import InputSocket, OutputSocket
+from haystack.core.component.component import Component, component
+from haystack.core.component.types import InputSocket, OutputSocket
 
 __all__ = ["component", "Component", "InputSocket", "OutputSocket"]
